App internal changes
Reorganized the app code to follow better structure practices
