+++ src/stonks_overwatch/app/__init__.py
@@ -1,0 +1,5 @@
+from .main import StonksOverwatchApp
+
+
+def main():
+    return StonksOverwatchApp("Stonks Overwatch", "com.caribay.stonks_overwatch")
